refactor: alterado quantidade de Long Para Double e adicionado uma validação de atributos especificos

+++ src/main/java/org/ong/pet/pex/backendpetx/dto/request/ProdutoDTO.java
@@ -21,7 +21,7 @@ public record ProdutoDTO(
         String descricao,
 
         @NotNull(message = "Preço do produto não pode ser nulo")
-        Long quantidade,
+        Double quantidade,
 
         @NotNull(message = "Preço do produto não pode ser nulo")
         UnidadeDeMedidaEnum unidadeDeMedida,
