+++ src/main/java/org/ong/pet/pex/backendpetx/dto/response/RacaoDisponivelResposta.java
@@ -4,7 +4,7 @@ import lombok.Builder;
 
 @Builder
 public record RacaoDisponivelResposta(
-        long quantidadeRacaoEmGramas,
+        double quantidadeRacaoEmGramas,
         double consumoDiraio,
         String racaoDias
 ) {
